+++ gui/src/App.tsx
@@ -1,6 +1,12 @@
-import { ThemeProvider } from "@mui/material";
+import {
+  PaletteMode,
+  ThemeProvider,
+  createTheme,
+  useMediaQuery,
+} from "@mui/material";
 import CssBaseline from "@mui/material/CssBaseline";
-import React from "react";
+import { grey } from "@mui/material/colors";
+import React, { useMemo } from "react";
 import { QueryClient, QueryClientProvider } from "react-query";
 import { Route, Router, Switch, useLocation } from "wouter";
 
@@ -8,14 +14,33 @@ import { HomePage } from "./components/HomePage";
 import { Navbar } from "./components/Navbar";
 import { TxReviewDialog } from "./components/TxReviewDialog";
 import { WagmiWrapper } from "./components/WagmiWrapper";
-import { useTheme } from "./hooks/useTheme";
 
 const queryClient = new QueryClient({
   defaultOptions: { queries: { suspense: true } },
 });
 
+const getDesignTokens = (mode: PaletteMode) => ({
+  palette: {
+    mode,
+    ...(mode === "light"
+      ? {
+          background: {
+            default: grey[50],
+          },
+        }
+      : {
+          background: {},
+        }),
+  },
+});
+
 export default function App() {
-  const theme = useTheme();
+  const prefersDarkMode = useMediaQuery("(prefers-color-scheme: dark)");
+
+  const theme = useMemo(
+    () => createTheme(getDesignTokens(prefersDarkMode ? "dark" : "light")),
+    [prefersDarkMode]
+  );
   const loc = useLocation();
   console.log(loc);
